Updated documentation and examples
Updated documentation and examples

+++ assemble.sh
@@ -3,5 +3,12 @@
 SCRIPT_PATH="$(realpath "${BASH_SOURCE[0]}")"
 SCRIPT_DIR="$(dirname "$SCRIPT_PATH")"
 
-rm $1.obj.gz
-java -cp "$SCRIPT_DIR"/lib/* Assembler $1.asm $2 $3 $4 $5 $6 $7 $8 $9
+if [[ -e "$1.obj.gz" ]]; then
+	rm $1.obj.gz
+fi
+if grep -qi '^[[:space:]]*__MAIN__:' $1.asm; then
+    hasMain="--hasMain"
+else
+    hasMain=""
+fi
+java -Dfile.encoding=UTF8 -cp "$SCRIPT_DIR"/lib/* Assembler $1.asm $hasMain $2 $3 $4 $5 $6 $7 $8 $9
