+++ run.sh
@@ -3,8 +3,13 @@
 SCRIPT_PATH="$(realpath "${BASH_SOURCE[0]}")"
 SCRIPT_DIR="$(dirname "$SCRIPT_PATH")"
 
-rm $1.obj.gz
-java -Dfile.encoding=UTF8 -cp "$SCRIPT_DIR"/lib/* Assembler $1.asm $2 $3 $4 $5 $6 $7 $8 $9
+MEMSIZE=0x100000
+STACKSIZE=0x2000
+
+if [[ -e "$1.obj.gz" ]]; then
+	rm $1.obj.gz
+fi
+java -Dfile.encoding=UTF8 -cp "$SCRIPT_DIR"/lib/* Assembler $1.asm
 if [[ -e "$1.obj.gz" ]] ; then
-	java -cp "$SCRIPT_DIR"/lib/* Simulation $1.obj.gz --verbose
+	java -Dfile.encoding=UTF8 -cp "$SCRIPT_DIR"/lib/* Simulation $1.obj.gz --verbose --mem=$MEMSIZE --stack=$STACKSIZE $2 $3 $4 $5 $6 $7 $8 $9
 fi
